Resolve circular dependencies
Proper separation of component and model logic
