+++ src/component/InspectorDetailsContent.tsx
@@ -8,7 +8,7 @@ import { ColumnDataPropType } from "./renderDataUtils";
 import { JsonSchemaGroup } from "../model/JsonSchemaGroup";
 import { createOptionTargetArrayFromIndexes, getFieldValueFromSchemaGroup } from "../model/schemaUtils";
 import { isDefined, listValues, commonValues, minimumValue, maximumValue } from "../model/utils";
-import { RenderColumn, RenderItemsColumn, RenderOptionsColumn } from "../types/Inspector";
+import { RenderColumn, RenderItemsColumn, RenderOptionsColumn } from "./InspectorTypes";
 import { TypeInRawJsonSchema, KeysOfRawJsonSchema } from "../types/RawJsonSchema";
 
 function containsTrueOrReduce<T>(
