+++ src/index.ts
@@ -3,7 +3,7 @@ import { getFieldValueFromSchemaGroup } from "./model/schemaUtils";
 import { minimumValue, maximumValue, commonValues, listValues } from "./model/utils";
 
 import { JsonSchemaGroup } from "./model/JsonSchemaGroup";
-import { KeysOfRawJsonSchema } from "./types/RawJsonSchema";
+import { KeysOfRawJsonSchema, KeysOfRawJsonSchemaWithValuesOf } from "./types/RawJsonSchema";
 
 /**
  * Main Inspector component (with numerous props).
@@ -22,7 +22,7 @@ export const Inspector = InspectorComponent;
  */
 export const getMinimumFieldValueFromSchemaGroup = (
     schemaGroup: JsonSchemaGroup,
-    fieldName: KeysOfRawJsonSchema,
+    fieldName: KeysOfRawJsonSchemaWithValuesOf<number>,
     defaultValue: number,
     optionIndexes?: Array<number>
 ): number => getFieldValueFromSchemaGroup(schemaGroup, fieldName, minimumValue, defaultValue, null, optionIndexes);
@@ -39,7 +39,7 @@ export const getMinimumFieldValueFromSchemaGroup = (
  */
 export const getMaximumFieldValueFromSchemaGroup = (
     schemaGroup: JsonSchemaGroup,
-    fieldName: KeysOfRawJsonSchema,
+    fieldName: KeysOfRawJsonSchemaWithValuesOf<number>,
     defaultValue: number,
     optionIndexes?: Array<number>
 ): number => getFieldValueFromSchemaGroup(schemaGroup, fieldName, maximumValue, defaultValue, null, optionIndexes);
