+++ src/types/RenderOptions.ts
@@ -1,0 +1,5 @@
+export type RenderOptions = {
+    groupTitle?: string;
+    options?: Array<RenderOptions>;
+    optionNameForIndex?: (indexes: Array<number>) => string | undefined;
+};
