+++ test/component/InspectorDetailsContent.test.tsx
@@ -5,7 +5,7 @@ import { InspectorDetailsContent, collectFormFields } from "../../src/component/
 import { JsonSchema } from "../../src/model/JsonSchema";
 import { createGroupFromSchema } from "../../src/model/schemaUtils";
 import { createRenderDataBuilder } from "../../src/component/renderDataUtils";
-import { RenderItemsColumn } from "../../src/types/Inspector";
+import { RenderItemsColumn } from "../../src/component/InspectorTypes";
 
 describe("renders correctly", () => {
     const buildColumnData = createRenderDataBuilder(() => (): void => {});
