+++ test/component/breadcrumbsUtils.test.ts
@@ -1,7 +1,7 @@
-import { createBreadcrumbBuilder } from "../../src/model/breadcrumbsUtils";
+import { createBreadcrumbBuilder } from "../../src/component/breadcrumbsUtils";
 import { JsonSchema } from "../../src/model/JsonSchema";
 import { JsonSchemaGroup } from "../../src/model/JsonSchemaGroup";
-import { RenderOptionsColumn, RenderItemsColumn } from "../../src/types/Inspector";
+import { RenderOptionsColumn, RenderItemsColumn } from "../../src/component/InspectorTypes";
 
 describe("createBreadcrumbBuilder()", () => {
     const simpleSchema = new JsonSchemaGroup().with(new JsonSchema({ title: "value" }, {}));
